intercom, clientmgr: stop retrying claims when appropriate
This introduces a function in intercom to stop claiming a client when it is
found to leave the node while the claim is not yet complete. During the first
4.5s when a client first joins the network there is a chance of the client
roaming away. This commit fixes the bug resulting in the client seemingly being
assigned to both nodes.

clientmgr_delete_client() now utilizes that function. Detecting removal of an
fdb entry or nl80211 neighbour will trigger the new behavior.